+++ dimas/src/com/message.rs
@@ -23,8 +23,8 @@ impl Deref for Message {
 
 impl Message {
 	/// decode message
-	/// # Errors
 	///
+	/// # Errors
 	pub fn decode<T>(self) -> Result<T>
 	where
 		T: for<'a> Decode<'a>,
@@ -34,7 +34,7 @@ impl Message {
 			.value
 			.try_into()
 			.map_err(|_| DimasError::ConvertingValue)?;
-		decode::<T>(value.as_slice()).map_err(|_| DimasError::DecodingMessage.into())
+		decode::<T>(value.as_slice()).map_err(|_| DimasError::Decoding.into())
 	}
 }
 // endregion:	--- Message
@@ -54,8 +54,8 @@ impl Deref for Request {
 
 impl Request {
 	/// Reply to the given request
-	/// # Errors
 	///
+	/// # Errors
 	#[allow(clippy::needless_pass_by_value)]
 	pub fn reply<T>(self, value: T) -> Result<()>
 	where
@@ -95,8 +95,8 @@ impl Deref for Response {
 
 impl Response {
 	/// decode response
-	/// # Errors
 	///
+	/// # Errors
 	pub fn decode<T>(self) -> Result<T>
 	where
 		T: for<'a> Decode<'a>,
@@ -106,10 +106,41 @@ impl Response {
 			.value
 			.try_into()
 			.map_err(|_| DimasError::ConvertingValue)?;
-		decode::<T>(value.as_slice()).map_err(|_| DimasError::DecodingMessage.into())
+		decode::<T>(value.as_slice()).map_err(|_| DimasError::Decoding.into())
 	}
 }
 // endregion:	--- Response
+
+// region:		--- Feedback
+/// Implementation of feedback messages
+#[derive(Debug)]
+pub struct Feedback(pub(crate) Sample);
+
+impl Deref for Feedback {
+	type Target = Sample;
+
+	fn deref(&self) -> &Self::Target {
+		&self.0
+	}
+}
+
+impl Feedback {
+	/// decode feedback
+	/// 
+	/// # Errors
+	pub fn decode<T>(self) -> Result<T>
+	where
+		T: for<'a> Decode<'a>,
+	{
+		let value: Vec<u8> = self
+			.0
+			.value
+			.try_into()
+			.map_err(|_| DimasError::ConvertingValue)?;
+		decode::<T>(value.as_slice()).map_err(|_| DimasError::Decoding.into())
+	}
+}
+// endregion:	--- Feedback
 
 #[cfg(test)]
 mod tests {
@@ -123,5 +154,6 @@ mod tests {
 		is_normal::<Message>();
 		is_normal::<Request>();
 		is_normal::<Response>();
+		is_normal::<Feedback>();
 	}
 }